add delete for export express

+++ functions.php
@@ -214,9 +214,16 @@ function delete_data_server($table_name, $ids,$user){
     foreach ($_ids as $id){
         if(isset($id)){
             $wp_table_name = get_wp_table_name($table_name);
+            $result = $wpdb->get_results($wp_table_name, "SELECT local_id FROM $wp_table_name WHERE id = '$id' AND user = '$user'");
+            if(!$result){
+                continue;
+            }
             $wpdb->delete($wp_table_name, array('id'=>$id));
-            $table_name = get_wp_table_name('delete_record');
-            $wpdb->insert($table_name,['table_name'=>$table_name,'record_id'=>$id,'user'=>$user]);
+            if($result[0]->local_id) {
+                $table_name = get_wp_table_name('delete_record');
+                $wpdb->insert($table_name,['table_name'=>$table_name,'record_id'=>$result[0]->local_id,'user'=>$user]);
+            }
+
         }
     }
